requirements.txt
langchain-community
openai
langchain_openai